STNDS-472: Create legal person Key Usage value validator (#37)
* STNDS-472: Create legal person Key Usage value validator
