+++ pkilint/etsi/en_319_412_2.py
@@ -1,17 +1,13 @@
-import enum
 import typing
-from typing import Optional
 
 from pyasn1.type import univ
 from pyasn1_alt_modules import rfc5280, rfc3739
 
 from pkilint import validation, oid, document, common
-from pkilint.etsi import asn1 as etsi_asn1
+from pkilint.etsi import asn1 as etsi_asn1, etsi_shared
 from pkilint.etsi import etsi_constants
 from pkilint.etsi.asn1 import en_319_411_2
-from pkilint.itu import bitstring
 from pkilint.pkix import extension, name, Rfc2119Word
-from pkilint.pkix.certificate.certificate_extension import KeyUsageBitName
 from pkilint.pkix.general_name import GeneralNameTypeName
 
 
@@ -197,111 +193,6 @@ class NaturalPersonExtensionIdentifierAllowanceValidator(common.ExtensionIdentif
             allowances[rfc3739.id_pe_qcStatements] = Rfc2119Word.MAY
 
         super().__init__(allowances, self._CODE_CLASSIFIER, Rfc2119Word.MAY)
-
-
-class KeyUsageValueValidator(validation.Validator):
-    """
-    NAT-4.3.2-1: The key usage extension shall be present and shall contain one (and only one) of the key usage settings
-    defined in table 1 (A, B, C, D, E or F).
-    """
-    VALIDATION_UNKNOWN_KEY_USAGE_SETTING = validation.ValidationFinding(
-        validation.ValidationFindingSeverity.ERROR,
-        'etsi.en_319_412_2.nat-4.3.2-1.unknown_key_usage_setting'
-    )
-
-    """
-    NAT-4.3.2-1: ... Type A, C or E should be used to avoid mixed usage of keys.
-    """
-    VALIDATION_MIXED_KEY_USAGE_SETTING = validation.ValidationFinding(
-        validation.ValidationFindingSeverity.WARNING,
-        'etsi.en_319_412_2.nat-4.3.2-1.mixed_key_usage_setting'
-    )
-
-    """
-    NAT-4.3.2-2: Certificates used to validate commitment to signed content (e.g. documents, agreements and/or
-    transactions) shall be limited to type A, B or F.
-    """
-    VALIDATION_INVALID_CONTENT_COMMITMENT_SETTING = validation.ValidationFinding(
-        validation.ValidationFindingSeverity.ERROR,
-        'etsi.en_319_412_2.nat-4.3.2-1.invalid_content_commitment_setting'
-    )
-
-    """
-    NAT-4.3.2-3: Of these alternatives, type A should be used (see the security note 2 below).
-    """
-    VALIDATION_NON_PREFERRED_CONTENT_COMMITMENT_SETTING = validation.ValidationFinding(
-        validation.ValidationFindingSeverity.WARNING,
-        'etsi.en_319_412_2.nat-4.3.2-3.non_preferred_content_commitment_setting'
-    )
-
-    _ALL_KUS = {str(n) for n in rfc5280.KeyUsage.namedValues}
-
-    def __init__(self, is_content_commitment_type: Optional[bool]):
-        super().__init__(
-            validations=[
-                self.VALIDATION_UNKNOWN_KEY_USAGE_SETTING,
-                self.VALIDATION_MIXED_KEY_USAGE_SETTING,
-                self.VALIDATION_INVALID_CONTENT_COMMITMENT_SETTING,
-                self.VALIDATION_NON_PREFERRED_CONTENT_COMMITMENT_SETTING,
-            ],
-            pdu_class=rfc5280.KeyUsage
-        )
-
-        self._is_content_commitment_type = is_content_commitment_type
-
-    class KeyUsageSetting(enum.Enum):
-        A = ({KeyUsageBitName.NON_REPUDIATION}, set())
-        B = ({KeyUsageBitName.NON_REPUDIATION, KeyUsageBitName.DIGITAL_SIGNATURE}, set())
-        C = ({KeyUsageBitName.DIGITAL_SIGNATURE}, set())
-        D = ({KeyUsageBitName.DIGITAL_SIGNATURE}, {KeyUsageBitName.KEY_AGREEMENT, KeyUsageBitName.KEY_ENCIPHERMENT})
-        E = (set(), {KeyUsageBitName.KEY_AGREEMENT, KeyUsageBitName.KEY_ENCIPHERMENT})
-        F = (
-            {KeyUsageBitName.NON_REPUDIATION, KeyUsageBitName.DIGITAL_SIGNATURE},
-            {KeyUsageBitName.KEY_AGREEMENT, KeyUsageBitName.KEY_ENCIPHERMENT}
-        )
-
-    _CONTENT_COMMITMENT_SETTINGS = {KeyUsageSetting.A, KeyUsageSetting.B, KeyUsageSetting.F}
-    _NON_CONTENT_COMMITMENT_SETTINGS = {s for s in KeyUsageSetting} - _CONTENT_COMMITMENT_SETTINGS
-
-    _MIXED_USE_SETTINGS = {KeyUsageSetting.B, KeyUsageSetting.D, KeyUsageSetting.F}
-
-    @classmethod
-    def _detect_setting(cls, key_usage_node: document.PDUNode) -> Optional[KeyUsageSetting]:
-        asserted_bits = {k for k in cls._ALL_KUS if bitstring.has_named_bit(key_usage_node, k)}
-
-        for setting in cls.KeyUsageSetting:
-            n_of_n_required_bits, one_of_n_required_bits = setting.value
-
-            allowed_bits = n_of_n_required_bits | one_of_n_required_bits
-
-            if (
-                    asserted_bits >= n_of_n_required_bits and
-                    (len(one_of_n_required_bits & asserted_bits) == 1 or not one_of_n_required_bits) and
-                    not any(asserted_bits - allowed_bits)
-            ):
-                return setting
-
-        return None
-
-    def validate(self, node):
-        setting = self._detect_setting(node)
-
-        if setting is None:
-            raise validation.ValidationFindingEncountered(self.VALIDATION_UNKNOWN_KEY_USAGE_SETTING)
-
-        if self._is_content_commitment_type is not None:
-            if self._is_content_commitment_type:
-                if setting not in self._CONTENT_COMMITMENT_SETTINGS:
-                    raise validation.ValidationFindingEncountered(self.VALIDATION_INVALID_CONTENT_COMMITMENT_SETTING)
-                elif setting != self.KeyUsageSetting.A:
-                    raise validation.ValidationFindingEncountered(
-                        self.VALIDATION_NON_PREFERRED_CONTENT_COMMITMENT_SETTING
-                    )
-            elif not self._is_content_commitment_type and setting not in self._NON_CONTENT_COMMITMENT_SETTINGS:
-                raise validation.ValidationFindingEncountered(self.VALIDATION_INVALID_CONTENT_COMMITMENT_SETTING)
-
-        if setting in self._MIXED_USE_SETTINGS:
-            raise validation.ValidationFindingEncountered(self.VALIDATION_MIXED_KEY_USAGE_SETTING)
 
 
 def _general_name_has_uri_prefixes(uri_prefixes_lower, general_name_node):
@@ -565,3 +456,29 @@ class CertificatePoliciesValidator(validation.Validator):
                 f'Certificate type is {self._certificate_type} ({self._recommended_policy_oid}) but certificate '
                 f'contains certificate type policy identifier "{policy_oid}"'
             )
+
+
+class NaturalPersonKeyUsageValidator(etsi_shared.KeyUsageValidator):
+    """
+    NAT-4.3.2-2: Certificates used to validate commitment to signed content (e.g. documents, agreements and/or
+    transactions) shall be limited to type A, B or F.
+    """
+    VALIDATION_INVALID_CONTENT_COMMITMENT_SETTING = validation.ValidationFinding(
+        validation.ValidationFindingSeverity.ERROR,
+        'etsi.en_319_412_2.nat-4.3.2-1.invalid_content_commitment_setting'
+    )
+
+    """
+    NAT-4.3.2-3: Of these alternatives, type A should be used (see the security note 2 below).
+    """
+    VALIDATION_NON_PREFERRED_CONTENT_COMMITMENT_SETTING = validation.ValidationFinding(
+        validation.ValidationFindingSeverity.WARNING,
+        'etsi.en_319_412_2.nat-4.3.2-3.non_preferred_content_commitment_setting'
+    )
+
+    def __init__(self, is_content_commitment_type):
+        super().__init__(
+            is_content_commitment_type,
+            self.VALIDATION_INVALID_CONTENT_COMMITMENT_SETTING,
+            self.VALIDATION_NON_PREFERRED_CONTENT_COMMITMENT_SETTING
+        )
