+++ tests/etsi/test_en_319_412_2.py
@@ -1,8 +1,9 @@
 import pytest
 from pyasn1_alt_modules import rfc5280
+
+import pkilint.etsi.etsi_shared
 from pkilint import document, validation
 from pkilint.etsi import en_319_412_2
-from pkilint.pkix.certificate import certificate_extension
 
 
 def _create_node(*bits):
@@ -12,7 +13,7 @@ def _create_node(*bits):
 
 
 def test_setting_a():
-    validator = en_319_412_2.KeyUsageValueValidator(None)
+    validator = en_319_412_2.NaturalPersonKeyUsageValidator(None)
 
     node = _create_node('nonRepudiation')
 
@@ -22,7 +23,7 @@ def test_setting_a():
 
 
 def test_setting_d():
-    validator = en_319_412_2.KeyUsageValueValidator(None)
+    validator = en_319_412_2.NaturalPersonKeyUsageValidator(None)
 
     node = _create_node('digitalSignature', 'keyEncipherment')
 
@@ -31,11 +32,11 @@ def test_setting_d():
     with pytest.raises(validation.ValidationFindingEncountered) as e:
         validator.validate(node)
 
-    assert e.value.finding == en_319_412_2.KeyUsageValueValidator.VALIDATION_MIXED_KEY_USAGE_SETTING
+    assert e.value.finding == pkilint.etsi.etsi_shared.KeyUsageValidator.VALIDATION_MIXED_KEY_USAGE_SETTING
 
 
 def test_setting_a_prohibited():
-    validator = en_319_412_2.KeyUsageValueValidator(False)
+    validator = en_319_412_2.NaturalPersonKeyUsageValidator(False)
 
     node = _create_node('nonRepudiation')
 
@@ -44,11 +45,14 @@ def test_setting_a_prohibited():
     with pytest.raises(validation.ValidationFindingEncountered) as e:
         validator.validate(node)
 
-    assert e.value.finding == en_319_412_2.KeyUsageValueValidator.VALIDATION_INVALID_CONTENT_COMMITMENT_SETTING
+    assert (
+            e.value.finding ==
+            en_319_412_2.NaturalPersonKeyUsageValidator.VALIDATION_INVALID_CONTENT_COMMITMENT_SETTING
+    )
 
 
 def test_setting_c_prohibited():
-    validator = en_319_412_2.KeyUsageValueValidator(True)
+    validator = en_319_412_2.NaturalPersonKeyUsageValidator(True)
 
     node = _create_node('digitalSignature')
 
@@ -57,11 +61,14 @@ def test_setting_c_prohibited():
     with pytest.raises(validation.ValidationFindingEncountered) as e:
         validator.validate(node)
 
-    assert e.value.finding == en_319_412_2.KeyUsageValueValidator.VALIDATION_INVALID_CONTENT_COMMITMENT_SETTING
+    assert (
+            e.value.finding ==
+            en_319_412_2.NaturalPersonKeyUsageValidator.VALIDATION_INVALID_CONTENT_COMMITMENT_SETTING
+    )
 
 
 def test_non_preferred_content_commitment_setting():
-    validator = en_319_412_2.KeyUsageValueValidator(True)
+    validator = en_319_412_2.NaturalPersonKeyUsageValidator(True)
 
     node = _create_node('digitalSignature', 'nonRepudiation')
 
@@ -70,11 +77,14 @@ def test_non_preferred_content_commitment_setting():
     with pytest.raises(validation.ValidationFindingEncountered) as e:
         validator.validate(node)
 
-    assert e.value.finding == en_319_412_2.KeyUsageValueValidator.VALIDATION_NON_PREFERRED_CONTENT_COMMITMENT_SETTING
+    assert (
+            e.value.finding ==
+            en_319_412_2.NaturalPersonKeyUsageValidator.VALIDATION_NON_PREFERRED_CONTENT_COMMITMENT_SETTING
+    )
 
 
 def test_mixed_use_setting():
-    validator = en_319_412_2.KeyUsageValueValidator(None)
+    validator = en_319_412_2.NaturalPersonKeyUsageValidator(None)
 
     node = _create_node('digitalSignature', 'nonRepudiation')
 
@@ -83,11 +93,11 @@ def test_mixed_use_setting():
     with pytest.raises(validation.ValidationFindingEncountered) as e:
         validator.validate(node)
 
-    assert e.value.finding == en_319_412_2.KeyUsageValueValidator.VALIDATION_MIXED_KEY_USAGE_SETTING
+    assert e.value.finding == pkilint.etsi.etsi_shared.KeyUsageValidator.VALIDATION_MIXED_KEY_USAGE_SETTING
 
 
 def test_invalid_extra_bit():
-    validator = en_319_412_2.KeyUsageValueValidator(None)
+    validator = en_319_412_2.NaturalPersonKeyUsageValidator(None)
 
     node = _create_node('digitalSignature', 'cRLSign')
 
@@ -96,11 +106,11 @@ def test_invalid_extra_bit():
     with pytest.raises(validation.ValidationFindingEncountered) as e:
         validator.validate(node)
 
-    assert e.value.finding == en_319_412_2.KeyUsageValueValidator.VALIDATION_UNKNOWN_KEY_USAGE_SETTING
+    assert e.value.finding == pkilint.etsi.etsi_shared.KeyUsageValidator.VALIDATION_UNKNOWN_KEY_USAGE_SETTING
 
 
 def test_invalid_both_bits():
-    validator = en_319_412_2.KeyUsageValueValidator(None)
+    validator = en_319_412_2.NaturalPersonKeyUsageValidator(None)
 
     node = _create_node('digitalSignature', 'keyAgreement', 'keyEncipherment')
 
@@ -109,4 +119,4 @@ def test_invalid_both_bits():
     with pytest.raises(validation.ValidationFindingEncountered) as e:
         validator.validate(node)
 
-    assert e.value.finding == en_319_412_2.KeyUsageValueValidator.VALIDATION_UNKNOWN_KEY_USAGE_SETTING
+    assert e.value.finding == pkilint.etsi.etsi_shared.KeyUsageValidator.VALIDATION_UNKNOWN_KEY_USAGE_SETTING
