+++ tests/integration_certificate/etsi/qevcp_w_eidas_pre_certificate/policy_critical.crttest
@@ -44,5 +44,5 @@ certificate.tbsCertificate.extensions.5.extnValue.certificatePolicies.1.policyQu
 certificate.tbsCertificate.extensions,SubscriberExtensionAllowanceValidator,WARNING,cabf.serverauth.subscriber.subject_key_identifier_extension_present,
 certificate.tbsCertificate.extensions,SubscriberExtensionAllowanceValidator,WARNING,cabf.serverauth.subscriber.unknown_extension_present,Unknown extension present: 1.3.6.1.5.5.7.1.3
 certificate.tbsCertificate.extensions.5,CertificatePoliciesCriticalityValidator,WARNING,etsi.en_319_412_2.gen-4.3.3-1.critical_certificate_policies_extension,Extension 2.5.29.32 is critical
-certificate.tbsCertificate.extensions.3.extnValue.keyUsage,KeyUsageValueValidator,WARNING,etsi.en_319_412_2.nat-4.3.2-1.mixed_key_usage_setting,
+certificate.tbsCertificate.extensions.3.extnValue.keyUsage,LegalPersonKeyUsageValidator,WARNING,etsi.en_319_412_2.nat-4.3.2-1.mixed_key_usage_setting,
 certificate.tbsCertificate.subject.rdnSequence,LegalPersonSubjectAttributeAllowanceValidator,ERROR,etsi.en_319_412_3.leg-4.2.1-2.organization_identifier_attribute_absent,
