+++ tests/integration_certificate/etsi/qevcp_w_psd2_final_certificate/invalid_orgid_scheme.crttest
@@ -70,4 +70,4 @@ certificate.tbsCertificate.subject.rdnSequence.3.0.value.x520OrganizationIdentif
 certificate.tbsCertificate.extensions.3.extnValue.certificatePolicies.2.policyQualifiers.0,CertificatePolicyQualifierValidator,WARNING,cabf.serverauth.certificate_policy_qualifier_present,
 certificate.tbsCertificate.subject.rdnSequence,EvSubscriberAttributeAllowanceValidator,WARNING,cabf.ev_guidelines.common_name_attribute_present,
 certificate.tbsCertificate.subject.rdnSequence.3.0.value.x520OrganizationIdentifier,PsdOrganizationIdentifierFormatValidator,ERROR,etsi.ts_119_495.gen-5.2.1-3.invalid_psd_organization_id_format,"Invalid PSD organization identifier format: ""PADFR-ACPR-30748"""
-certificate.tbsCertificate.extensions.8.extnValue.keyUsage,KeyUsageValueValidator,WARNING,etsi.en_319_412_2.nat-4.3.2-1.mixed_key_usage_setting,
+certificate.tbsCertificate.extensions.8.extnValue.keyUsage,LegalPersonKeyUsageValidator,WARNING,etsi.en_319_412_2.nat-4.3.2-1.mixed_key_usage_setting,
