+++ tests/integration_certificate/etsi/qevcp_w_psd2_pre_certificate/semantics_info_nra_missing.crttest
@@ -62,4 +62,4 @@ certificate.tbsCertificate.extensions.8.extnValue.keyUsage,SubscriberKeyUsageVal
 certificate.tbsCertificate.extensions,SubscriberExtensionAllowanceValidator,WARNING,cabf.serverauth.subscriber.unknown_extension_present,Unknown extension present: 1.3.6.1.5.5.7.1.3
 certificate.tbsCertificate.extensions,SubscriberExtensionAllowanceValidator,WARNING,cabf.serverauth.subscriber.unknown_extension_present,Unknown extension present: 2.23.140.3.1
 certificate.tbsCertificate.subject.rdnSequence.3.0.value.x520OrganizationIdentifier,PsdOrganizationIdentifierFormatValidator,ERROR,etsi.ts_119_495.gen-5.2.1-3.invalid_psd_organization_id_format,"Invalid PSD organization identifier format: ""XX:GR-18050"""
-certificate.tbsCertificate.extensions.8.extnValue.keyUsage,KeyUsageValueValidator,WARNING,etsi.en_319_412_2.nat-4.3.2-1.mixed_key_usage_setting,
+certificate.tbsCertificate.extensions.8.extnValue.keyUsage,LegalPersonKeyUsageValidator,WARNING,etsi.en_319_412_2.nat-4.3.2-1.mixed_key_usage_setting,
