+++ tests/integration_certificate/etsi/qncp_w_iv_eidas_final_certificate/semantics_info_nra_uri_missing.crttest
@@ -55,5 +55,5 @@ certificate.tbsCertificate.extensions.5.extnValue.certificatePolicies.0.policyQu
 certificate.tbsCertificate.extensions.9.extnValue.qCStatements.4.statementInfo.semanticsInformation,NaturalPersonIdentifierNameRegistrationAuthoritiesValidator,ERROR,etsi.en_319_412_1.nat-5.1.3-05.national_id_scheme_name_registration_authorities_uri_missing,
 certificate.tbsCertificate.extensions,SubscriberExtensionAllowanceValidator,WARNING,cabf.serverauth.subscriber.subject_key_identifier_extension_present,
 certificate.tbsCertificate.extensions,SubscriberExtensionAllowanceValidator,WARNING,cabf.serverauth.subscriber.unknown_extension_present,Unknown extension present: 1.3.6.1.5.5.7.1.3
-certificate.tbsCertificate.extensions.2.extnValue.keyUsage,KeyUsageValueValidator,WARNING,etsi.en_319_412_2.nat-4.3.2-1.mixed_key_usage_setting,
+certificate.tbsCertificate.extensions.2.extnValue.keyUsage,NaturalPersonKeyUsageValidator,WARNING,etsi.en_319_412_2.nat-4.3.2-1.mixed_key_usage_setting,
 certificate.tbsCertificate.extensions.5.extnValue.certificatePolicies,CertificatePoliciesValidator,ERROR,etsi.en_319_412_2.qcs-5.2-2.mismatched_policy_identifier_for_certificate_type,"Certificate type is QNCP_W_IV_EIDAS_FINAL_CERTIFICATE (0.4.0.194112.1.5) but certificate contains certificate type policy identifier ""0.4.0.194112.1.4"""
